Move statistics files to one folder
