+++ frontend/smart-kickers-game/src/components/Game/GameStatistics/StatisticsItems/FastestShot.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './GameStatistics.css';
+import './StatisticItem.css';
 
 function FastestShot({ blue, white, statistics }) {
   function returnFastestShot(teamID) {
