--- frontend/smart-kickers-game/src/components/Game/GameStatistics/StatisticsItems/StatisticItem.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import './StatisticItem.css';
-
-function StatisticsItem({ children }) {
-  return <div className="table-item">{children}</div>;
-}
-
-export default StatisticsItem;
